requirements.txt
redis==5.0.1
httpx==0.27.0
python-dotenv==1.0.0
pydantic==2.5.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
aiofiles==23.2.1
qdrant-client==1.7.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
streamlit==1.31.0

# Phoenix observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-instrumentation-httpx==0.42b0

# Expositional Engineering dependencies
networkx==3.2.1
numpy==1.26.0

# Skills framework
pyyaml==6.0.1

# Trigger.dev (optional - for cloud long-running tasks)
# trigger-sdk==3.0.0

